+++ ublox/src/parser.rs
@@ -920,8 +920,8 @@ mod test {
     fn parser_oom_clears_buffer() {
         let bytes = CfgNav5Builder {
             mask: CfgNav5Params::DYN,
-            dyn_model: CfgNav5DynModel::AirborneWithLess1gAcceleration,
-            fix_mode: CfgNav5FixMode::Only3D,
+            dyn_model: NavDynamicModel::AirborneWithLess1gAcceleration,
+            fix_mode: NavFixMode::Only3D,
             fixed_alt: 100.17,
             fixed_alt_var: 0.0017,
             min_elev_degrees: 17,
@@ -970,8 +970,8 @@ mod test {
     fn parser_accepts_packet_array_underlying() {
         let bytes = CfgNav5Builder {
             mask: CfgNav5Params::DYN,
-            dyn_model: CfgNav5DynModel::AirborneWithLess1gAcceleration,
-            fix_mode: CfgNav5FixMode::Only3D,
+            dyn_model: NavDynamicModel::AirborneWithLess1gAcceleration,
+            fix_mode: NavFixMode::Only3D,
             fixed_alt: 100.17,
             fixed_alt_var: 0.0017,
             min_elev_degrees: 17,
@@ -1002,8 +1002,8 @@ mod test {
     fn parser_accepts_packet_vec_underlying() {
         let bytes = CfgNav5Builder {
             mask: CfgNav5Params::DYN,
-            dyn_model: CfgNav5DynModel::AirborneWithLess1gAcceleration,
-            fix_mode: CfgNav5FixMode::Only3D,
+            dyn_model: NavDynamicModel::AirborneWithLess1gAcceleration,
+            fix_mode: NavFixMode::Only3D,
             fixed_alt: 100.17,
             fixed_alt_var: 0.0017,
             min_elev_degrees: 17,